+++ website/src/components/Install/Installer/StarterMini/index.tsx
@@ -16,6 +16,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 import { CopyIcon } from "@/assets";
+import { useAxelixVersion } from "@/hooks/useAxelixVersion";
 
 import { useRef, useState } from "react";
 
@@ -30,7 +31,8 @@ export const StarterMini = ({ artifact, activeSnippetRef }: IProps) => {
     const [copiedIdx, setCopiedIdx] = useState<number | null>(null);
     const timers = useRef<Array<ReturnType<typeof setTimeout> | null>>([null, null, null]);
 
-    const axelixActualVersion = `1.0.0`;
+    const { version } = useAxelixVersion();
+    const axelixActualVersion = version ?? "1.0.0";
 
     const blocks: { label: string; code: string }[] = [
         {
