+++ website/src/components/Install/Installer/index.tsx
@@ -16,14 +16,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 "use client";
-import {
-    EInstallConfigurationVariant,
-    EInstallMethod,
-    EInstallOpenSelect,
-    ESpringBootVariant,
-    IAxelixVersionData,
-    IGithubReleaseResponseBody,
-} from "@/models";
+import { useAxelixVersion } from "@/hooks/useAxelixVersion";
+import { EInstallConfigurationVariant, EInstallMethod, EInstallOpenSelect, ESpringBootVariant } from "@/models";
 import { installSpringBootArtifact } from "@/utils";
 
 import { BareMetal } from "./Snippets/BareMetal";
@@ -53,10 +47,7 @@ export const Installer = () => {
     const activeSnippetRef = useRef<HTMLElement>(null);
 
     const [openSelect, setOpenSelect] = useState<EInstallOpenSelect>(EInstallOpenSelect.NULL);
-    const [axelixVersionData, setAxelixVersionData] = useState<IAxelixVersionData>({
-        version: null,
-        loading: true,
-    });
+    const axelixVersionData = useAxelixVersion();
 
     const selectRef = useRef<HTMLDivElement>(null);
 
@@ -75,43 +66,6 @@ export const Installer = () => {
     }, []);
 
     // TODO: In the future, split this component into smaller components
-    useEffect(() => {
-        async function fetchAxelixVersion() {
-            try {
-                setAxelixVersionData((prev) => ({
-                    ...prev,
-                    loading: true,
-                }));
-
-                const response = await fetch("https://api.github.com/repos/axelixlabs/axelix/releases/latest");
-
-                if (!response.ok) {
-                    throw new Error();
-                }
-
-                const data: IGithubReleaseResponseBody = await response.json();
-                const version = data.tag_name;
-
-                setAxelixVersionData((prev) => ({
-                    ...prev,
-                    version: version,
-                }));
-            } catch {
-                setAxelixVersionData((prev) => ({
-                    ...prev,
-                    version: "1.0.0",
-                }));
-            } finally {
-                setAxelixVersionData((prev) => ({
-                    ...prev,
-                    loading: false,
-                }));
-            }
-        }
-
-        fetchAxelixVersion();
-    }, []);
-
     return (
         <>
             <div className={styles.InstallerWrapper}>
